% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA__onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Standard onLoad function for rcreds}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{libname and pkgname are default \code{R} \code{.onLoad()} parameters}

\item{pkgname}{libname and pkgname are default \code{R} \code{.onLoad()} parameters}
}
\value{
invisible(TRUE)

(not exported)
}
\description{
Sets several options, if they are not already set when this function funs.
}
