% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_functions.r,
%   R/credentials_functions_DB_specific.r
\name{writeCredentialsToFile}
\alias{writeCredentialsToFile}
\alias{readCredentialsFromFile}
\alias{read_write_redentials_files}
\alias{readDBCredentialsFromFile}
\alias{writeDBCredentialsToFile}
\title{Read to / Write from credential files}
\usage{
writeCredentialsToFile(..., file_full_path = "auto", info.file_name = "",
  file_name = getOption("rcreds.file_name", default = ".credentials.creds"),
  folder = getOption("rcreds.folder", default = "~/.rcreds/credential_files"),
  zArchive_existing = TRUE, overwrite_existing = FALSE,
  key = readKeyFromFile(), verbose = getOption("verbose.rcreds", default =
  TRUE))

readCredentialsFromFile(file_full_path = "auto", info.file_name = "",
  file_name = getOption("rcreds.file_name", default = ".credentials.creds"),
  folder = getOption("rcreds.folder", default = "~/.rcreds/credential_files"),
  key = readKeyFromFile())

readDBCredentialsFromFile(file_full_path = "auto", info.file_name = "",
  file_name = getOption("rcreds.db.file_name", default =
  ".db_credentials.creds"), folder = getOption("rcreds.db.folder", default =
  "~/.rcreds/db_credential_files"), key = readKeyFromFile())

writeDBCredentialsToFile(dbname = "dev", host = "localhost", port = 5432,
  username = "you_forgot_to_specify_username",
  password = "too_many_secrets", file_full_path = "auto",
  info.file_name = "", file_name = getOption("rcreds.db.file_name", default
  = ".db_credentials.creds"), folder = getOption("rcreds.db.folder", default =
  "~/.rcreds/db_credential_files"), zArchive_existing = TRUE,
  overwrite_existing = FALSE, key = readKeyFromFile(), ...)
}
\arguments{
\item{TBD}{}
}
\value{
for \code{writeCredentialsToFile} & \code{writeDBCredentialsToFile}: The file path where the encrypted values have been stored. 

for \code{readCredentialsFromFile} & \code{readDBCredentialsFromFile}: A list object of the decrypted values.
}
\description{
Securely write sensitive info to disk and then read in and decrypt.
}
\details{
There are two sets of pairs of functions
use \code{writeCredentialsToFile()} to output to disk
use \code{readCredentialsFromFile()} to read in the credentials back to R

Similarly, there are a pair of functions with the 5 comonly-used parameters for database connections
use \code{writeDBCredentialsToFile()} and \code{readDBCredentialsFromFile()}
}
\examples{
  "TBD"
}
