% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_rcreds_file_full_path.R
\name{construct_rcreds_file_full_path}
\alias{construct_rcreds_file_full_path}
\title{Construct Rcreds File Full Path}
\usage{
construct_rcreds_file_full_path(file_name = getOption("rcreds.file_name",
  default = ".credentials.creds"), folder = get_default_rcreds_folder(DB =
  DB), info.file_name = "", DB = FALSE)
}
\arguments{
\item{file_name}{string, cannot be empty  Filename with extension. Default is '.credentials.creds'}

\item{folder}{string, cannot be empty  Where file will be stored}

\item{info.file_name}{string, can    be empty. Prepended to file_name.}

\item{DB}{TRUE/FALSE flag. Is the folder for the db credentials functions?

          Defaults to: FALSE}
}
\value{
The full file path
}
\description{
Takes parts of a file name, info, and folder and creates a full path
}
