% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_functions.r,
%   R/credentials_functions_DB_specific.r
\name{write_credentials_to_file}
\alias{write_credentials_to_file}
\alias{read_credentials_from_file}
\alias{read_write_redentials_files}
\alias{read_db_credentials_from_file}
\alias{write_db_credentials_to_file}
\title{Read to / Write from credential files}
\usage{
write_credentials_to_file(..., file_full_path = "auto", info.file_name = "",
  file_name = getOption("rcreds.file_name", default = ".credentials.creds"),
  folder = getOption("rcreds.folder", default = "~/.rcreds/credential_files"),
  zArchive_existing = TRUE, overwrite_existing = FALSE,
  key = readKeyFromFile(), verbose = getOption("verbose.rcreds", default =
  TRUE))

read_credentials_from_file(file_full_path = "auto", info.file_name = "",
  file_name = getOption("rcreds.file_name", default = ".credentials.creds"),
  folder = getOption("rcreds.folder", default = "~/.rcreds/credential_files"),
  key = readKeyFromFile())

read_db_credentials_from_file(file_full_path = "auto", info.file_name = "",
  file_name = getOption("rcreds.db.file_name", default =
  ".db_credentials.creds"), folder = getOption("rcreds.db.folder", default =
  "~/.rcreds/db_credential_files"), key = readKeyFromFile())

write_db_credentials_to_file(dbname = "dev", host = "localhost",
  port = 5432, username = "you_forgot_to_specify_username",
  password = "too_many_secrets", file_full_path = "auto",
  info.file_name = "", file_name = getOption("rcreds.db.file_name", default
  = ".db_credentials.creds"), folder = getOption("rcreds.db.folder", default =
  "~/.rcreds/db_credential_files"), zArchive_existing = TRUE,
  overwrite_existing = FALSE, key = readKeyFromFile(), ...)
}
\arguments{
\item{file_full_path}{The pathname}
}
\value{
for \code{write_credentials_to_file} & \code{write_db_credentials_to_file}: The file path where the encrypted values have been stored. 

for \code{read_credentials_from_file} & \code{read_db_credentials_from_file}: A list object of the decrypted values.
}
\description{
Securely write sensitive info to disk and then read in and decrypt.
}
\details{
There are two sets of pairs of functions
use \code{write_credentials_to_file()} to output to disk
use \code{read_credentials_from_file()} to read in the credentials back to R

Similarly, there are a pair of functions with the 5 comonly-used parameters for database connections
use \code{write_db_credentials_to_file()} and \code{read_db_credentials_from_file()}
}
\examples{

  some_login_function <- function(username, password) {
    ## does something with username/password
    ## ... 
  }

  credentials_list <- read_db_credentials_from_file()
  some_login_function(username = credentials_list$user_name
                    , password = credentials_list$password
                     )

}
